Refactored class locations.
