+++ spring-boot/admin/src/main/java/eu/coatrack/admin/service/ProxyDockerComposeTemplateInitializationFailedException.java
@@ -1,4 +1,4 @@
-package eu.coatrack.admin.controllers;
+package eu.coatrack.admin.service;
 
 public class ProxyDockerComposeTemplateInitializationFailedException extends RuntimeException {
     public ProxyDockerComposeTemplateInitializationFailedException(Exception cause) {
